Fix: Add loading spinner and update app icon
A loading spinner has been added to the `QuizScreen` to indicate that questions are being loaded. The `QuizScreen` now displays a `CircularProgressIndicator` while `questionSet` is null.

The app icon has been updated. The following changes were made:
- New foreground and background assets for the launcher icon have been added.
- The `ic_launcher.xml` and `ic_launcher_round.xml` files have been updated to use the new assets.
- A new color resource `ic_launcher_background.xml` has been added.
- A Play Store icon `ic_launcher-playstore.png` has been added.

The `MainActivity` has also been updated to wrap the `QuizScreen` with a `Surface` composable, utilizing the `MaterialTheme.colorScheme.background`.